chore: Challenge 페이지 단독 사용 파일 pages/Challenge 로 이동
