+++ src/shared/interface/atomsType.ts
@@ -19,10 +19,6 @@ export interface InputProps extends themeProps {
   description?: string;
   maxLength: number;
 }
-export interface ResultProps {
-  result: 'success' | 'fail' | 'progress';
-  index?: number;
-}
 
 export interface TitleProps {
   children: React.ReactNode;
@@ -34,32 +30,6 @@ export interface ButtonProps extends themeProps {
   cancel?: boolean;
   event?: MouseEventHandler<HTMLButtonElement>;
   type?: 'button' | 'submit' | 'reset';
-}
-
-export interface ChallengeProps {
-  idx?: number;
-  title: string;
-  memo: string;
-  day: number;
-  result: 'success' | 'fail' | 'progress';
-  index?: number;
-  successCheck: boolean;
-}
-
-export interface NewChallengeProps {
-  title: string;
-  memo: string;
-  day: number;
-  result: 'progress';
-  successCheck: boolean;
-}
-
-//TODO - 서버 연결시 삭제하기
-export interface DummyProps {
-  dummy: ChallengeProps[];
-  setDummy: (challenge: ChallengeProps) => void;
-  deleteDummy: (index: number) => void;
-  updateDummy: (index: number, result?: 'progress' | 'success' | 'fail', successCheck?: boolean) => void;
 }
 
 export interface SlideProps {
